peerspace-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
passlib[bcrypt]
bcrypt
python-jose
email-validator
google-genai
python-dotenv